Move team list fragment to common graphql folder
