+++ services/manage/app/graphql/fragments/team-list.js
