+++ services/manage/app/routes/manage/orgs/view/apps/view/teams.js
@@ -1,7 +1,7 @@
 import Route from '@ember/routing/route';
 import AppQueryMixin from '@base-cms/id-me-manage/mixins/app-query';
 import gql from 'graphql-tag';
-import fragment from './teams/list-fragment';
+import fragment from '@base-cms/id-me-manage/graphql/fragments/team-list';
 
 const query = gql`
   query AppTeams {
